+++ CHANGELOG.md
@@ -5,6 +5,9 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 ## [Unreleased]
 
+###Added
+- Output of consumer name so that it's easy to see where the DLQ message was going to
+
 ## [3.1.0] - 2018-10-30
 
 ### Added